cilium: convert to common.mk (#153)
* cilium: convert to common.mk

* overrides targets for now

+++ build/lib/create_release_checksums.sh
@@ -39,11 +39,13 @@ for file in $(find ${ASSET_ROOT} -type f -not -path '*\.sha[25][51][62]' -not -p
     sha512sum "$filepath" | tee -a $SHA512SUM > "$file.sha512" || return 1
 done
 
-mv $SHA256SUM $SHA512SUM "$ASSET_ROOT"
-sha256sum SHA256SUM > "SHA256SUM.sha256" || return 1
-sha512sum SHA256SUM > "SHA256SUM.sha512" || return 1
-sha256sum SHA512SUM > "SHA512SUM.sha256" || return 1
-sha512sum SHA512SUM > "SHA512SUM.sha512" || return 1
+if [ -f $SHA256SUM ]; then
+    mv $SHA256SUM $SHA512SUM "$ASSET_ROOT"
+    sha256sum SHA256SUM > "SHA256SUM.sha256" || return 1
+    sha512sum SHA256SUM > "SHA256SUM.sha512" || return 1
+    sha256sum SHA512SUM > "SHA512SUM.sha256" || return 1
+    sha512sum SHA512SUM > "SHA512SUM.sha512" || return 1
 
-cat SHA256SUM
+    cat SHA256SUM
+fi
 cd -
